Problem 22
Solution using file containing first names (e.g. included names.txt) to sort file in alphabetical order and calculate the value of each name like the following:

For example, when the list is sorted into alphabetical order, COLIN, which is worth 3 + 15 + 12 + 9 + 14 = 53, is the 938th name in the list. So, COLIN would obtain a score of 938 × 53 = 49714.